source/geh_common/pyproject.toml
[project]
name = "geh_common"
version = "5.5.0"
description = "Functionality common to DataHub3 subsystems"
readme = "README.md"

requires-python = ">=3.11"
dependencies = [
    "pyspark>=3.5.0",
    "delta-spark>=3.1,<4.0",
    "dependency_injector>=4.43.0,<5.0",
    "azure-monitor-opentelemetry>=1.6.0",
    "azure-core>=1.30.0,<2.0.0",
    "azure-identity>=1.16.1,<2.0.0",
    "azure-keyvault-secrets>=4.7.0,<5.0.0",
    "azure-monitor-query>=1.2.0,<2.0.0",
    "pyyaml>=6.0.2",
    "databricks-sdk>=0.42.0",
    "pydantic-settings>=2.7.1",
    "pydantic>=2.10.6",
    "polars>=1.25.2", # will be removed, just temporary to create the version which still creates csv files
    "streamlit>=1.44.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "psutil>=6.1.1",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.6.1",
    "ruff>=0.9.4",
]

[tool.pytest.ini_options]
# -s ensures that print("<text>") is written to output.
addopts = [
    "--cov-report=term",
    "--cov-report=xml:coverage.xml",
    "--cov=geh_common",
    "-ra",                           # Show extra test summary info as specified by -ra
    "-vv",                           # Verbose output for tests
    "-s",                            # Print to console
    "-n=logical",                    # Run tests in parallel
    "--dist=worksteal",              # Pop tests off the queue in chunks
]
cache_dir = ".cache/pytest"

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]

[tool.ruff]
line-length = 120
show-fixes = true

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["D", "D401", "E", "F", "I", "T201", "UP"]
ignore = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D106",
    "D107",
    "D417",
    "E501",
    "UP006",
    "UP007",
    "UP035",
]
exclude = ["notebooks/*"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.uv.workspace]
members = ["source/geh_common"]
